refactor: Migrate multiplayer page to use game engine
Before: 600+ lines of mixed logic and UI
After: 150 lines using useGameEngine hook

Changes:
- Replaced manual state management with StateManager
- Replaced scattered fetch calls with NetworkSync
- Replaced polling logic with engine.start()
- Removed 450+ lines of boilerplate code

Benefits:
- 75% code reduction
- Cleaner separation of concerns
- Easier to test and maintain
- Reusable for tournaments/AI modes

Backup: page-old.tsx kept for reference

Tests: 30 passing ✅
